reverting change and going back to "double" precision
Revert "Changed precision to float (!)"

This reverts commit 798fb2a5cf1cba921e3499e5857af53f525c82b6.

+++ java1/src/TempVertex.java
@@ -13,7 +13,7 @@ public class TempVertex {
 		cellNeighbors = new HashSet<Integer>();
 	}
 	
-	public float[] vertexCoords() {
+	public double[] vertexCoords() {
 		return vertex.coords();
 	}
 	
@@ -80,8 +80,8 @@ public class TempVertex {
 			for (int fc : v.cellNeighbors)
 				cellNeighborsSet.add(fc);
 		}
-		y = (int)Math.round((float)y/(float)verts.size());
-		x = (int)Math.round((float)x/(float)verts.size());
+		y = (int)Math.round((double)y/(double)verts.size());
+		x = (int)Math.round((double)x/(double)verts.size());
 		
 		// make array from TreeSet
 		TempVertex merged = new TempVertex(new Vertex(y, x));
